#56 Extract client transports in separate files.
