+++ lib/registry.js
@@ -209,7 +209,7 @@ class ServiceRegistry {
 				examples,
 				snippets,
 				code,
-				client: fs.readFileSync(__dirname + '/../lib/client.js')
+				client: fs.readFileSync(__dirname + '../lib/client/index.js')
 			}, { _with: false }, function (err, file) {
 				res.send(err ? err.toString() : file);
 			});
@@ -253,8 +253,8 @@ class ServiceRegistry {
 					}
 				});
 			}
-		} else if (requestUrl == '/client' || requestUrl == '/client.js') {
-			fs.readFile(__dirname + '/../lib/client.js', {encoding: 'utf-8'}, function (err, text) {
+		} else if (requestUrl == '/client' || requestUrl == '/index.js') {
+			fs.readFile(__dirname + '../lib/index.js', {encoding: 'utf-8'}, function (err, text) {
 				if (err) res.status(404).send('Error reading client library\n\n' + err);
 				else {
 					res.set('Content-Type', 'application/javascript');
