+++ test/runtest.bat
@@ -1,3 +1,3 @@
 @echo off
 cmd /c mocha --ui tdd --reporter spec server.js
-cmd /c mocha --ui tdd --reporter spec client.js
+cmd /c mocha --ui tdd --reporter spec index.js
